Use PAM to change passwords
Modifying /etc/shadow directly can be wrong in many configurations,
and will fail to update other tokens (such as keyring passwords).
Instead, use a small helper to interact with PAM (as PAM functions
are not safe to use in process). Communication with the helper
happens through a pipe, which also solves the problem of exposing
the crypted password to all users via /proc.

As the PAM stack requires an uncrypted password, the API of SetPassword
was changed, and now the password travels in plaintext across the
system bus. A later patch will solve this security issue using
reversible encryption.

https://bugs.freedesktop.org/show_bug.cgi?id=51833