daemon: Create our directories in /var on demand
For OSTree, I'm trying to move daemons to fully create and manage
their local state (/var) at runtime.  The benefit of this is clarity;
right now it's a weird hybrid where rpm/dpkg create empty directories,
and then the daemon populates them further.

Having this be at runtime means that we don't need to deal with
uid/gid allocation on the build side.