lib: ignore closing sessions
The function act_user_is_logged_in() determines the logged-in-ness
of a user by looking at all the session for that user. Recently,
systemd introduced a session state 'closing' for session which are
'nominally logged out, but not quite gone yet'. These sessions
should be ignored for the purpose of determining logged-in-ness.
If we don't, then logging out of a GNOME session shows the user
as still logged in on the login screen, and refuses to let him
choose another session.