lib: fix set_has_multiple_users call in remove_uesr
The remove_user() function removes a user from the
various caches and then if the cache size drops to
one sets the has-multiple-users property to FALSE.

Well, that's what it's supposed to do. It actually
sets it to FALSE if cache size is greater than 1.

This commit fixes that.

Spotted by Florian Müllner