ActUserManagerNewSession: do not free if there are pending async calls
While the ActUserManagerNewSession is the user_data for an asynchronous
call, it is not safe to free the struct: when the async call completes,
it will dereference the struct.

In GIO, cancellation happens asynchronously, so using a GCancellable is
not enough to prevent this: the cancelled async call still won't complete
until we get back to the main loop.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=50112
Bug-Debian: http://bugs.debian.org/681826