lib: map session to user objects not user names
A user may not be fully loaded when a session comes in.
In this case, we create a user object and start loading it,
then map the session object to the username of the user object
(which may not be available yet).

This commit drops the username indirection, and maps sessions
straight to user objects.