Use GcrSecretExchange to pass the password from the client to the daemon
GcrSecretExchange, from gcr, provides a way for two programs to communicate
secret data without the risk of eavesdropping.
We use it to pass the password, which would otherwise travel in plain-text,
from the UI to the daemon. The two step protocol is implemented by two
methods, BeginSetPassword and ContinueSetPassword.
Due to the nature of the PIN implementation, and given that password and
hint were tied before, ContinueSetPassword is actually a setter for all
those three at once.
For backwards compatibility, SetPassword is retained and continues to
take the plain-text password, but it is deprecated. SetPIN is removed
because it never appeared in a stable release.

https://bugs.freedesktop.org/show_bug.cgi?id=51833