Avoid an error when generating enums in distcheck
We were passing the full filename of all prerequisites to
glib-mkenums, but we were also switching to srcdir - this works
as long as switching to srcdir does not change our level in the
directory tree. With distcheck it does, and the filenames of the
form ../../../src/libaccountsservice/foo.h point to the wrong
place. Fix this by passing just the file-within-directory part
of the prerequisites.