user: never treat cached users as system accounts
Right now the check for "is system user" is less than solid.
It depends on checking things like whether or not a password is set,
and whether or not a shell is set. Unfortunately, sometimes a password
is not set for a non-system account, and sometimes a shell is set for
a system account, so the heuristics aren't perfect.

In one particular case, we currently get things wrong: right after a
user is created before they have a password.

This commit attempts to shore up the heuristics by never treating cached
users as system users.

https://bugs.freedesktop.org/show_bug.cgi?id=64769