ActUserManager: handle nonexistent users
Right now if a nonexistent user is requested from the accounts service,
we don't give the caller any sort of notication that the user doesn't
exist. This cascades into d-bus timeouts and ultimately a broken login
screen.

This commit adds a new "nonexistent" user property to the user object
and sets it to TRUE when a requested user fails to load.  It also
makes sure the user objects is-loaded property gets set.

Based on a patch by Lionel Landwerlin <llandwerlin@gmail.com>

https://bugs.freedesktop.org/show_bug.cgi?id=66325